Implement Coinbase Advanced Trade client
