+++ src/__init__.py
@@ -1,0 +1,1 @@
+from .spot_swap_hedge import *  # noqa: F401,F403
